+++ lib/siridb/siridb.py
@@ -112,12 +112,16 @@ async def query_all_series_results(siridb_client, series_name, selector="*"):
     return result
 
 
-async def query_series_forecasts(siridb_client, series_name, selector="*"):
+async def query_series_forecasts(siridb_client, series_name, selector="*",
+                                 only_future=False):
     result = None
+    after = ""
+    if only_future:
+        after = " after now"
     try:
         result = await siridb_client.query(
             f'select {selector} from '
-            f'/enodo_{re.escape(series_name)}_forecast_.*?$/')
+            f'/enodo_{re.escape(series_name)}_forecast_.*?$/{after}')
     except (QueryError, InsertError, ServerError, PoolError,
             AuthenticationError, UserAuthError) as e:
         print("Connection problem with SiriDB server")