+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - Endpoint for fetching all analysis results for a series (for all job types)
+- Query param `future` for the `api/series/anomaly_test_series/forecasts` endpoint, to only return points in the future
 
 ## [0.1.0-beta3.2.12] - 2022-05-23